feat(commercetools:api-client): split the index.ts into graphqlError.ts, graphqlFactory.ts and authHelpers.ts
